requirements.txt
# Core ML dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0

# Feature extraction
fairseq>=0.12.0
praat-parselmouth>=0.4.0

# UI
gradio>=4.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
omegaconf>=2.3.0

# Optional but recommended
accelerate>=0.20.0
